requirements.txt
requests>=2.31.0
anthropic>=0.40.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
rich>=13.0.0
streamlit>=1.30.0
brightdata-sdk>=2.2.0
google-genai>=1.0.0
